% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{add_new_experiment}
\alias{add_new_experiment}
\title{Create a new model tracking experiment}
\usage{
add_new_experiment(name)
}
\arguments{
\item{name}{your experiment name - the experiment sub-directory within
the tracking folder will be given this name}
}
\description{
Creates a new model tracking experiment with the given
name, including sub-directories to store model parameters, metrics and
artifacts
}
\examples{
\\notrun{
add_new_experiment(name = "new_experiment")
}
}
