% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{check_tracking_folder}
\alias{check_tracking_folder}
\title{creates a tracking folder in the root directory if doesn't
already exist}
\usage{
check_tracking_folder()
}
\description{
creates a tracking folder in the root directory if doesn't
already exist
}
