% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{create_folder}
\alias{create_folder}
\title{helper function to create a new folder in the
given location if it doesn't already exist}
\usage{
create_folder(location)
}
\description{
helper function to create a new folder in the
given location if it doesn't already exist
}
