% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logartifact.R
\name{log_artifact}
\alias{log_artifact}
\title{Log an artifact}
\usage{
log_artifact(x, experiment, i)
}
\arguments{
\item{x}{the object to save}

\item{experiment}{the name of the current experiment}

\item{i}{run number}
}
\description{
Save an \code{R} object created during an
experiment run
}
\note{
currently supported objects that can be written are
\itemize{
  \item \code{ggplot2}
  \item \code{data.frame}
  \item \code{matrix}
  \item \code{list}
  \item \code{vector}
}
}
