% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_file}
\alias{log_file}
\title{Write a file associated with the current experiment to files}
\usage{
log_file(experiment, filelocation)
}
\description{
Write a file associated with the current experiment to files
}
