% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_metric}
\alias{log_metric}
\title{Log the value of a model evaluation metric}
\usage{
log_metric(experiment, metric, as)
}
\arguments{
\item{experiment}{name of the current experiment}

\item{metric}{the metric to write out}

\item{as}{name to call the saved metric}
}
\description{
Writes the value of the provided metric to a file
in the metrics sub-directory of the current experiment
}
\note{
It is recommended not to change the value of \code{as} from one run
of the experiment to the next as it will not be possible to evaluate the
changing performance of the model with the changing value of \code{metric}
}
\examples{
\dontrun{
log_param("experiment", metric = rmse, as = "rmse")
}
}
