% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_param}
\alias{log_param}
\title{Log the value of a model parameter}
\usage{
log_param(experiment, param, as)
}
\arguments{
\item{experiment}{name of the current experiment}

\item{param}{the parameter to write out}

\item{as}{name to call the saved parameter}
}
\description{
Writes the value of the provided parameter to a file
in the params sub-directory of the current experiment
}
\note{
It is recommended not to change the value of \code{as} from one run
of the experiment to the next as it will not be possible to evaluate the
changing performance of the model with the changing value of \code{param}
}
\examples{
\dontrun{
x <- 10
log_param("experiment", param = x, as = "x_value")
}
}
